// frontend/modules/shared/types/buttonTypes.ts
export enum ButtonStyle {
  Primary = 'Primary',
  Outline = 'Outline',
  Tab = 'Tab',
  NoBackground = 'NoBackground',
  Cancel = 'Cancel',
  FormSubmit = 'FormSubmit'
}

export enum ButtonSize {
  Minimal = 'Minimal',
  Small = 'Small',
  Medium = 'Medium',
  Large = 'Large'
}

export enum ButtonType {
  Button = 'button',
  Submit = 'submit'
}